feat: player controls, example audio and ui fixes
